+++ spotify/manager/session.py
@@ -23,6 +23,13 @@ class SpotifySessionManager(object):
           the re-login feature. The blob is obtained from the
           :meth:`credentials_blob_updated` callback after a successful
           login to the Spotify AP.
+
+    When behind a proxy, the application can specify:
+        - `proxy`: url to the proxy server that should be used. The format
+            is ``protocol://<host>:port`` (where protocol is
+            ``http/https/socks4/socks5``)
+        - `proxy_username`: username to authenticate with the proxy server.
+        - `proxy_password`: password to authenticate with the proxy server.
     """
 
     cache_location = 'tmp'
@@ -32,7 +39,8 @@ class SpotifySessionManager(object):
     user_agent = 'pyspotify-example'
 
     def __init__(self, username=None, password=None, remember_me=False,
-                 login_blob=''):
+                 login_blob='', proxy=None, proxy_username=None,
+                 proxy_password=None):
         self._cmdqueue = Queue.Queue()
 
         # Session settings
@@ -48,6 +56,9 @@ class SpotifySessionManager(object):
         self.username = username
         self.password = password
         self.remember_me = remember_me
+        self.proxy = proxy
+        self.proxy_username = proxy_username
+        self.proxy_password = proxy_password
         self.login_blob = login_blob
 
         # Create session