+++ src/session.c
@@ -902,6 +902,7 @@ create_session(Session *self, PyObject *client, PyObject *settings)
     sp_session *session;
     sp_error error;
     char *cache_location, *settings_location, *user_agent;
+    char *proxy, *proxy_username, *proxy_password;
 
     memset(&config, 0, sizeof(config));
     config.api_version = SPOTIFY_API_VERSION;
@@ -957,6 +958,38 @@ create_session(Session *self, PyObject *client, PyObject *settings)
     fprintf(stderr, "[DEBUG]-session- User agent set to '%s'\n",
                         user_agent);
 #endif
+
+    proxy = PySpotify_GetConfigString(client, "proxy", 1);
+    if (!proxy)
+        return -1;
+    if ((long) proxy != (-1)) {
+    config.proxy = proxy;
+#ifdef DEBUG
+    fprintf(stderr, "[DEBUG]-session- Proxy set to '%s'\n", proxy);
+#endif
+    }
+
+    proxy_username = PySpotify_GetConfigString(client, "proxy_username", 1);
+    if (!proxy_username)
+       return -1;
+    if ((long) proxy_username != (-1)) {
+        config.proxy_username = proxy_username;
+#ifdef DEBUG
+        fprintf(stderr, "[DEBUG]-session- Proxy username set to '%s'\n",
+            proxy_username);
+#endif
+    }
+
+    proxy_password = PySpotify_GetConfigString(client, "proxy_password", 1);
+    if (!proxy_password)
+       return -1;
+    if ((long) proxy_password != (-1)) {
+        config.proxy_password = proxy_password;
+#ifdef DEBUG
+        fprintf(stderr, "[DEBUG]-session- Proxy password set to '%s'\n",
+            proxy_password);
+#endif
+    }
 
 #ifdef DEBUG
     fprintf(stderr, "[DEBUG]-session- creating session...\n");